:recycle: split regression tests
